students/K3341/Alekseev_Andrey/lr_1/requirements.txt
fastapi[all]>=0.115.0
sqlmodel>=0.0.22
alembic>=1.13.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9

pydantic